Move dialect files
