+++ dpcomp/CMakeLists.txt
@@ -37,7 +37,7 @@ if (NOT EXPECTED_LLVM_SHA STREQUAL LLVM_SHA)
     message(FATAL_ERROR "Invalid llvm version")
 endif()
 
-add_subdirectory(include/plier)
+add_subdirectory(include/plier/dialect/plier)
 
 set(SOURCES_LIST
     lib/Conversion/SCFToAffine/SCFToAffine.cpp
@@ -45,7 +45,7 @@ set(SOURCES_LIST
     lib/analysis/memory_ssa_analysis.cpp
     lib/compiler/compiler.cpp
     lib/compiler/pipeline_registry.cpp
-    lib/dialect.cpp
+    lib/dialect/plier/dialect.cpp
     lib/transforms/arg_lowering.cpp
     lib/transforms/call_lowering.cpp
     lib/transforms/canonicalize_reductions.cpp
@@ -70,12 +70,12 @@ set(SOURCES_LIST
     )
 set(HEADERS_LIST
     include/plier/Conversion/SCFToAffine/SCFToAffine.h
-    include/plier/PlierOps.td
     include/plier/analysis/memory_ssa.hpp
     include/plier/analysis/memory_ssa_analysis.hpp
     include/plier/compiler/compiler.hpp
     include/plier/compiler/pipeline_registry.hpp
-    include/plier/dialect.hpp
+    include/plier/dialect/plier/PlierOps.td
+    include/plier/dialect/plier/dialect.hpp
     include/plier/transforms/arg_lowering.hpp
     include/plier/transforms/call_lowering.hpp
     include/plier/transforms/canonicalize_reductions.hpp
