+++ dpcomp/include/plier/transforms/call_lowering.hpp
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "plier/dialect.hpp"
+#include "plier/dialect/plier/dialect.hpp"
 
 #include <mlir/IR/PatternMatch.h>
