+++ numba_dpcomp/mlir_compiler/lib/pipelines/plier_to_std.cpp
@@ -32,7 +32,7 @@
 #include <mlir/Transforms/GreedyPatternRewriteDriver.h>
 #include <mlir/Transforms/Passes.h>
 
-#include "plier/dialect.hpp"
+#include "plier/dialect/plier/dialect.hpp"
 
 #include "plier/transforms/call_lowering.hpp"
 #include "plier/transforms/cast_utils.hpp"
