+++ numba_dpcomp/mlir_compiler/lib/py_map_types.cpp
@@ -19,7 +19,7 @@
 #include <mlir/IR/BuiltinTypes.h>
 #include <mlir/IR/TypeRange.h>
 
-#include "plier/dialect.hpp"
+#include "plier/dialect/plier/dialect.hpp"
 
 namespace py = pybind11;
